requirements.txt
faiss-cpu
sentence-transformers
openai
python-dotenv
PyPDF2
tiktoken
streamlit
google-generativeai